[otel-contrib] Add opentelemetry-contri library.
* Move b3 propagator and jaeger propagator into opentelemetry-contrib library.

* Move composite propagator's test into opentelemetry-contrib library since it has more propagators available.

* Modify the example in composite_propagator since we no longer have b3 propagator and jaeger propagator in main library.

* In order to avoid cyclic dependencies, opentelemetry-contrib is not one of the opentelemetry's dependencies.
