+++ opentelemetry-contrib/CHANGELOG.md
@@ -1,0 +1,7 @@
+# Changelog
+
+## v0.1.0
+
+### Added
+- B3 propagator(moved from api module)
+- Jaeger propagator
