+++ opentelemetry-contrib/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "opentelemetry-contrib"
+version = "0.1.0"
+authors = ["OpenTelemetry Authors <cncf-opentelemetry-contributors@lists.cncf.io>"]
+description = "Rust contrib repo for OpenTelemetry"
+homepage = "https://github.com/open-telemetry/opentelemetry-rust/tree/master/opentelemetry-contrib"
+repository = "https://github.com/open-telemetry/opentelemetry-rust/tree/master/opentelemetry-contrib"
+readme = "README.md"
+categories = [
+    "development-tools::debugging",
+    "development-tools::profiling",
+]
+keywords = ["opentelemetry","tracing"]
+license = "Apache-2.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+opentelemetry = { version = "0.8.0", path = ".." }
+lazy_static = "1.4"
