+++ opentelemetry-contrib/README.md
@@ -1,0 +1,3 @@
+# OpenTelemetry Contrib
+
+Contrib repo for OpenTelemetry Rust implementation.
