+++ opentelemetry-contrib/src/lib.rs
@@ -1,0 +1,8 @@
+//! # OpenTelemetry Contrib
+//!
+//! This is a library for extensions that is not part of core api. But still may be found useful by
+//! some users.
+//!
+//! Typically, those include vendor specific propagators
+
+pub mod trace_propagator;
