+++ opentelemetry-contrib/src/trace_propagator/mod.rs
@@ -1,0 +1,9 @@
+//! Vendor specific context propagator
+//!
+//!
+
+mod b3_propagator;
+mod jaeger_propagator;
+
+pub use b3_propagator::{B3Encoding, B3Propagator};
+pub use jaeger_propagator::JaegerPropagator;
