+++ src/api/trace/mod.rs
@@ -110,12 +110,10 @@
 //! Please review the W3C specification for details on the [Tracestate
 //! field](https://www.w3.org/TR/trace-context/#tracestate-field).
 //!
-pub mod b3_propagator;
 pub mod context;
 pub mod event;
 pub mod futures;
 pub mod id_generator;
-pub mod jaeger_propagator;
 pub mod link;
 pub mod noop;
 pub mod provider;
